chore: change aead aes gcm nonce from uint64_t to uint16_t
API and test cases modified to reflect changes.

+++ implementations/c/lib/vault/impl.h
@@ -170,7 +170,7 @@ typedef struct {
    */
   ockam_error_t (*aead_aes_gcm_encrypt)(ockam_vault_t*        vault,
                                         ockam_vault_secret_t* key,
-                                        uint64_t              nonce,
+                                        uint16_t              nonce,
                                         const uint8_t*        additional_data,
                                         size_t                additional_data_length,
                                         const uint8_t*        plaintext,
@@ -195,7 +195,7 @@ typedef struct {
    */
   ockam_error_t (*aead_aes_gcm_decrypt)(ockam_vault_t*        vault,
                                         ockam_vault_secret_t* key,
-                                        uint64_t              nonce,
+                                        uint16_t              nonce,
                                         const uint8_t*        additional_data,
                                         size_t                additional_data_length,
                                         const uint8_t*        ciphertext_and_tag,
